% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bed_data}
\alias{bed_data}
\title{Chromatin state predictions from Varshney, et al 2017}
\format{A list with one data frame with 53 rows and 5 variables:
\describe{
  \item{V1}{Chromosome}
  \item{V2}{genomic start coordinate}
  \item{V3}{genomic end coordinate}
  \item{V4}{chromatin state label}
  \item{V5}{chromatin state color for plotting}
  ...
}}
\usage{
bed_data
}
\description{
A list of data frames from bed files with genomic regions and chromatin state predictions
}
\keyword{datasets}
