% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_chr.R
\name{fix_chr}
\alias{fix_chr}
\title{Fix chromosome encoding}
\usage{
fix_chr(chr1, chr2)
}
\arguments{
\item{chr1}{vector of reference chromosome names}

\item{chr2}{chromosome encodings to change}
}
\description{
Fix chromosome encoding
}
