% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bed_track.R
\name{make_bed_track}
\alias{make_bed_track}
\title{Make a track for bed files}
\usage{
make_bed_track(bed_data, title = "", genome_build = "hg19",
  background_color = "#485467", background_frame = T)
}
\arguments{
\item{bed_data}{List of dataframes in bed format: chr\t start\t end\t label\t color}

\item{title}{title for sidebar}

\item{genome_build}{hg19 or hg38}

\item{background_color}{List of background colors for bed plot title bars}

\item{background_frame}{Should frames be added around each bed plot?}
}
\value{
\code{bed.track} gviz track
}
\description{
Make a track for bed files
}
