% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gene_track.R
\name{make_gene_track}
\alias{make_gene_track}
\title{Make a mart object for plotting}
\usage{
make_gene_track(chr, start, end, mart, genome_build = "hg19",
  highlight = NULL, title = "Ensembl", background_color = "#E69F00",
  background_frame = T)
}
\arguments{
\item{chr}{interval chromosome}

\item{start}{interval start}

\item{end}{interval end}

\item{mart}{mart object}

\item{genome_build}{either hg19 or hg38}

\item{highlight}{what gene to higlight}

\item{title}{title for sidebar}

\item{background_color}{color of sidebar}

\item{background_frame}{whether to add frame}
}
\value{
\code{gene.track} gviz track
}
\description{
Make a mart object for plotting
}
