% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ld_legend.R
\name{make_ld_legend}
\alias{make_ld_legend}
\title{Generate a regional plot from summary statistics for a desired interval}
\usage{
make_ld_legend(box.x = 4, box.ymin = 3.6, box.ymax = 4.6,
  box.width = NULL, box.colors = c("#B8B8B8", "#357EBD", "#46B8DA",
  "#5CB85C", "#EEA236", "#D43F3A"), tic.width = 0.1,
  tic.text.offset.x = 0.05, tic.text.offset.y = 0, tic.text = c("NA",
  "0.0", "0.2", "0.4", "0.6", "0.8", "1.0"))
}
\arguments{
\item{box.x}{x position of legend}

\item{box.ymin}{minimum y positions of legend}

\item{box.ymax}{maximum y positions of legend}

\item{box.width}{width of legend}

\item{box.colors}{vector of colors, one for each box in the legend}

\item{tic.width}{tic width}

\item{tic.text.offset.x}{x distance to offset tic label text}

\item{tic.text.offset.y}{y distance to offset tic label text}

\item{tic.text}{text to display next to each tic}
}
\description{
Generate a regional plot from summary statistics for a desired interval
}
