% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mart.R
\name{make_mart}
\alias{make_mart}
\title{Make a mart object for plotting}
\usage{
make_mart(genome_build = "hg19")
}
\arguments{
\item{genome_build}{either hg19 or hg38}
}
\value{
\code{mart} mart object
}
\description{
Make a mart object for plotting
}
