% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_regional_plot.R
\name{make_regional_plot}
\alias{make_regional_plot}
\title{Generate a regional plot from summary statistics for a desired interval}
\usage{
make_regional_plot(chr, start, end, variant_data, variant_chr_column,
  variant_pos_column, variant_y_column, variant_marker_column = NULL,
  variant_ld_data = NULL, variant_ld_ref = NULL,
  genome_build = "hg19", variant_should_log = T,
  variant_horizontal_value = 5e-08, variant_horizontal_color = "red",
  variant_horizontal_style = "dashed", variant_background_color = NULL,
  variant_background_frame = T, variant_point_color = "#000000",
  variant_title = "P-value", gene_highlight = NULL,
  gene_title = "Ensembl", gene_background_color = NULL,
  gene_frame = T, bed_data = NULL, bed_titles = NULL,
  bed_background_colors = NULL, bed_frame = T)
}
\arguments{
\item{chr}{interval chromosome}

\item{start}{interval start}

\item{end}{interval end}

\item{variant_data}{variant-level summary statistics}

\item{variant_chr_column}{chromosome column in \code{variant_data}}

\item{variant_pos_column}{variant position column in \code{variant_data}}

\item{variant_y_column}{p-value column in \code{variant_data}}

\item{variant_marker_column}{unique variant identifier column in \code{variant_data}}

\item{variant_ld_data}{variant LD values, the output of LDGds workflow}

\item{variant_ld_ref}{reference variant unique identifier for LD}

\item{genome_build}{genome build of variants}

\item{variant_should_log}{should the pvalues in \code{variant_data} be -log10'ed?}

\item{variant_horizontal_value}{Y-intercept for horizontal line in Manhattan plot}

\item{variant_horizontal_color}{Color for horizontal line in Manhattan plot}

\item{variant_horizontal_style}{Style for horizontal line in Manhattan plot}

\item{variant_background_color}{Manhattan plot title box color}

\item{variant_background_frame}{Should a frame be placed around the Manhattan plot?}

\item{variant_point_color}{Color of points if no LD information is provided}

\item{variant_title}{Manhattan plot title}

\item{gene_highlight}{Should a gene be highlighted? hgnc symbol (DEPRICATED)}

\item{gene_title}{Title for gene plot}

\item{gene_background_color}{Gene plot title box color}

\item{gene_frame}{Should a frame be placed around the Gene plot?}

\item{bed_data}{List of dataframes in bed format: chr start end label color}

\item{bed_titles}{List of titles for bed plots}

\item{bed_background_colors}{List of background colors for bed plot title bars}

\item{bed_frame}{Should frames be added around each bed plot?}
}
\description{
Generate a regional plot from summary statistics for a desired interval
}
