% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_variant_track.R
\name{make_variant_track}
\alias{make_variant_track}
\title{Create a manhattan plot}
\usage{
make_variant_track(variant_data, chr_column, pos_column, y_column,
  marker_column = NULL, ld_data = NULL, ld_ref = NULL,
  genome_build = "hg19", should_log = T, horizontal_value = 5e-08,
  horizontal_color = "red", horizontal_style = "dashed",
  background_color = "#e6550d", background_frame = T,
  point_color = "#252525", title = " ")
}
\arguments{
\item{variant_data}{variant-level summary statistics}

\item{chr_column}{chromosome column in \code{variant_data}}

\item{pos_column}{variant position column in \code{variant_data}}

\item{y_column}{p-value column in \code{variant_data}}

\item{marker_column}{unique variant identifier column in \code{variant_data}}

\item{ld_data}{variant LD values, the output of LDGds workflow}

\item{ld_ref}{reference variant unique identifier for LD}

\item{genome_build}{genome build of variants}

\item{should_log}{should the pvalues in \code{variant_data} be -log10'ed?}

\item{horizontal_value}{Y-intercept for horizontal line in Manhattan plot}

\item{horizontal_color}{Color for horizontal line in Manhattan plot}

\item{horizontal_style}{Style for horizontal line in Manhattan plot}

\item{background_color}{Manhattan plot title box color}

\item{background_frame}{Should a frame be placed around the Manhattan plot?}

\item{point_color}{Color of points if no LD information is provided}

\item{title}{Manhattan plot title}

\item{chr}{interval chromosome}

\item{start}{interval start}

\item{end}{interval end}
}
\value{
\code{variant.track} gviz track
}
\description{
Create a manhattan plot
}
