% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variant_data}
\alias{variant_data}
\title{Mock summary statistics from 1kg and fake phenotypes}
\format{A data frame with 426 rows and 13 variables:
\describe{
  \item{MarkerName}{unique variant identifier, shared with LD data}
  \item{chr}{chromosome of the variant}
  \item{pos}{genomic position of the variant}
  \item{ref}{reference allele}
  \item{alt}{alternate allele}
  \item{pvalue}{pvalue from GWAS}
  ...
}}
\usage{
variant_data
}
\description{
Variant level summary statistics as output from the Analysis Commons Genesis_wdl workflow
diamonds.
}
\keyword{datasets}
