% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variant_ld_data}
\alias{variant_ld_data}
\title{LD r^2 values for a small number of variants from 1kg phase 3}
\format{A data frame with 1 row and 115 variables:}
\usage{
variant_ld_data
}
\description{
LD r^2 values for a small number of variants from 1kg phase 3
}
\keyword{datasets}
